Fixed issue in create_variant() when passing an empty year range variable to filter().
When filter() is called with the geography code and a variable representing the years before npp, if the year range variable was empty but not None, filter() would not function properly and would return an Error. This was fixed in an earlier commit, however the previous fix still resulted in a number of Warnings that did not mean anything. The new change will (hopefully) remove all these warnings, and may remove the need for the earlier fix, but this needs to be tested.